# perf: optimize instr sci call query