--- a/apps/frontend/src/graphql/call/getCallsByInstrumentScientist.graphql
+++ b/apps/frontend/src/graphql/call/getCallsByInstrumentScientist.graphql
@@ -1,5 +1,7 @@
 query getCallsByInstrumentScientist($scientistId: Int!) {
   callsByInstrumentScientist(scientistId: $scientistId) {
-    ...call
+    id
+    shortCode
+    templateId
   }
 }