Closes: UserOfficeProject/issue-tracker#1250

## Description
This PR optimizes the query for retrieving calls by Instrument Scientist. 

Comparing the new query against the old on STFC prod:

| Old | New |
| ------------- | ------------- |
| 2.404904s | 0.127819s | 
| 2.129852s | 0.112924s | 
| 1.974664s | 0.119315s | 
| 3.013225s | 0.109058s | 
| 3.062653s | 0.107943s | 
| 2.059708s | 0.105477s | 
| 2.124660s | 0.109718s | 
| 4.493919s | 0.133059s | 
| 2.233179s | 0.146521s | 
| 3.068420s | 0.194828s | 



## Motivation and Context
The change is required to improve the performance of the application. It solves the problem of unnecessary data being fetched and processed, which slows down the application.

## Changes
- The calls prop in `CallFilterProps` and `ProposalFilterBarProps` has been modified to only pick 'shortCode' and 'id' fields.
- The GraphQL query `getCallsByInstrumentScientist` has been altered to return only 'id', 'shortCode', and 'templateId'.
- A new type `InstrumentScientistCallsData` has been introduced in `useInstrumentScientistCallsData.ts` to reflect the changes.
- The state `calls` in `useInstrumentScientistCallsData.ts` now uses the new type `InstrumentScientistCallsData`.

These changes ensure that only necessary data is fetched and processed, improving the application's performance.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Fixes Jira Issue

[https://jira.esss.lu.se/browse/](https://jira.esss.lu.se/browse/)

## Depends On

<!--- Does this PR depend on another PR that should be merged first or at the same time -->

## Tests included/Docs Updated?

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->

- [ ] I have added tests to cover my changes.
- [ ] All relevant doc has been updated